Move component monitoring to its own package
Agent monitoring has two separate functions - implementing the control
plane monitoring server and self-monitoring for components. Having both
in the same packages caused a dependency cycle involving the otel
translation package. Resolve this by putting component monitoring in a
subpackage
