+++ internal/pkg/agent/application/monitoring/component/testdata/monitoring_config_full.yaml
